+++ threedi_api_client/aio/openapi/api_client.py
@@ -2,7 +2,7 @@
 """
     3Di API
 
-    3Di simulation API (latest stable version: v3)   Framework release: 3.4.81   3Di core release: 3.6.14  deployed on:  02:33PM (UTC) on November 12, 2025  # noqa: E501
+    3Di simulation API (latest stable version: v3)   Framework release: 3.4.81.1   3Di core release: 3.6.15  deployed on:  09:37AM (UTC) on November 13, 2025  # noqa: E501
 
     The version of the OpenAPI document: v3
     Contact: info@nelen-schuurmans.nl